refactor: make a mwu running
