+++ jina/helloworld/chatbot/app.py
@@ -1,12 +1,13 @@
 import os
+import urllib.request
 import webbrowser
 from pathlib import Path
 
 from jina import Flow, Document
 from jina.importer import ImportExtensions
 from jina.logging import default_logger
+from jina.logging.profile import ProgressBar
 from jina.parsers.helloworld import set_hw_chatbot_parser
-from .helper import download_data
 
 
 def hello_world(args):
@@ -68,6 +69,30 @@ def hello_world(args):
             f.block()
 
 
+def download_data(targets, download_proxy=None, task_name='download fashion-mnist'):
+    """
+    Download data.
+
+    :param targets: target path for data.
+    :param download_proxy: download proxy (e.g. 'http', 'https')
+    :param task_name: name of the task
+    """
+    opener = urllib.request.build_opener()
+    opener.addheaders = [('User-agent', 'Mozilla/5.0')]
+    if download_proxy:
+        proxy = urllib.request.ProxyHandler(
+            {'http': download_proxy, 'https': download_proxy}
+        )
+        opener.add_handler(proxy)
+    urllib.request.install_opener(opener)
+    with ProgressBar(task_name=task_name, batch_unit='') as t:
+        for k, v in targets.items():
+            if not os.path.exists(v['filename']):
+                urllib.request.urlretrieve(
+                    v['url'], v['filename'], reporthook=lambda *x: t.update_tick(0.01)
+                )
+
+
 if __name__ == '__main__':
-    args = set_hw_chatbot_parser().parse_args([])
+    args = set_hw_chatbot_parser().parse_args()
     hello_world(args)
