--- jina/helloworld/chatbot/helper.py
@@ -1,28 +1,0 @@
-import os
-import urllib.request
-
-from jina.logging.profile import ProgressBar
-
-
-def download_data(targets, download_proxy=None, task_name='download fashion-mnist'):
-    """
-    Download data.
-
-    :param targets: target path for data.
-    :param download_proxy: download proxy (e.g. 'http', 'https')
-    :param task_name: name of the task
-    """
-    opener = urllib.request.build_opener()
-    opener.addheaders = [('User-agent', 'Mozilla/5.0')]
-    if download_proxy:
-        proxy = urllib.request.ProxyHandler(
-            {'http': download_proxy, 'https': download_proxy}
-        )
-        opener.add_handler(proxy)
-    urllib.request.install_opener(opener)
-    with ProgressBar(task_name=task_name, batch_unit='') as t:
-        for k, v in targets.items():
-            if not os.path.exists(v['filename']):
-                urllib.request.urlretrieve(
-                    v['url'], v['filename'], reporthook=lambda *x: t.update_tick(0.01)
-                )
