+++ jina/types/arrays/__init__.py
@@ -1,4 +1,3 @@
 from .chunk import ChunkArray
 from .document import DocumentArray
 from .match import MatchArray
-from .querylang import QueryLangArray
